saga/reducer re-routed, and delete working but not rendering until page refresh
