+++ src/redux/sagas/sites.saga.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import { put, takeLatest } from "redux-saga/effects";
 
 //SITES POST ROUTE
-function* addSite() {
+function* addSite(action) {
     try {
       //POST request to sites router
       const response = yield axios.post(`/api/sites/`, action.payload);
@@ -18,7 +18,7 @@ function* addSite() {
   }
   
   //SITES PUT ROUTE
-  function* updateSite() {
+  function* updateSite(action) {
       try {
         //UPDATE request sent to sites.router based on ID
         const response = yield axios.put(`/api/sites/${action.payload}`);
@@ -37,14 +37,14 @@ function* addSite() {
     try{
       //Get All sites for populating explore page
       const response = yield axios.get('api/sites');
-      yield put({type: 'SET_LIST', payload: response.data});
+      yield put({type: 'SET_ADMIN_SITES', payload: response.data});
     }catch(err){
       yield put({type: 'FETCH_SITES_ERR'})
     }
   }
   
   //SITES DELETE ROUTE
-  function* deleteSite() {
+  function* deleteSite(action) {
       try {
         //DELETE request sent to sites.router based on ID
         const response = yield axios.delete(`/api/sites/${action.payload}`);
